requirements.txt
yfinance>=0.2.28
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0
ta>=0.10.2
plotly>=5.14.0
python-dateutil>=2.8.2
pytz>=2023.3
tqdm>=4.65.0